requirements.txt
Django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
pyTelegramBotAPI==4.9.0
python-telegram-bot==20.0
celery==5.2.7
redis==4.4.2
psycopg2_binary==2.9.5